Готово!

+++ src/main/java/com/martin/service/BookService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface BookService {
 
-    Book add(String title, int authorId, int genreId) throws Exception;
+    Book add(String title, long authorId, long genreId) throws Exception;
     Book add(String authorTitle, String authorFirsname, String authorLastname, String genreTitle) throws Exception;
     long getCount();
     List<Book> getAll(int page, int amountByOnePage);
@@ -15,7 +15,7 @@ public interface BookService {
     Book findById(long id) throws Exception;
     List<Comment> getComments(long id);
     Book update(long id, String title) throws Exception;
-    Book update(long id, String bookTitle, int authorId, int genreId) throws Exception;
+    Book update(long id, String bookTitle, long authorId, long genreId) throws Exception;
     void delete(long id) throws Exception;
     void deleteAll();
 }
